+++ src/test/java/study/StationRepositoryTest.java
@@ -7,10 +7,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import study.oneway.NoGeneratedIdStation;
-import study.oneway.NoGeneratedIdStationRepository;
-import study.oneway.Station;
-import study.oneway.StationRepository;
+import study.manytoone.NoGeneratedIdStation;
+import study.manytoone.NoGeneratedIdStationRepository;
+import study.manytoone.Station;
+import study.manytoone.StationRepository;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DataJpaTest
