+++ src/test/java/study/manytoone/StationLineRelationTest.java
@@ -1,4 +1,4 @@
-package study.oneway;
+package study.manytoone;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
